parser started being created

+++ formulas/area.go
@@ -1,246 +1,246 @@
 package unitconv
 
-// AcretoHectare (acre float32) float32 in Hectare value
-func AcretoHectare(acre float32) float32 {
+// AcretoHectare (acre float64) float64 in Hectare value
+func AcretoHectare(acre float64) float64 {
 	return acre / 2.471
 }
 
-// AcretoSquareinch (acre float32) float32 in Square inch value
-func AcretoSquareinch(acre float32) float32 {
+// AcretoSquareinch (acre float64) float64 in Square inch value
+func AcretoSquareinch(acre float64) float64 {
 	return acre * 6.273e+6
 }
 
-// AcretoSquarefoot (acre float32) float32 in Square Foot value
-func AcretoSquarefoot(acre float32) float32 {
+// AcretoSquarefoot (acre float64) float64 in Square Foot value
+func AcretoSquarefoot(acre float64) float64 {
 	return acre * 43560
 }
 
-// AcretoSquareyard (acre float32) float32 in Square yard value
-func AcretoSquareyard(acre float32) float32 {
+// AcretoSquareyard (acre float64) float64 in Square yard value
+func AcretoSquareyard(acre float64) float64 {
 	return acre * 4840
 }
 
-// AcretoSquaremile (acre float32) float32 in Square mile
-func AcretoSquaremile(acre float32) float32 {
+// AcretoSquaremile (acre float64) float64 in Square mile
+func AcretoSquaremile(acre float64) float64 {
 	return acre / 640
 }
 
-// AcretoSquaremeter (acre float32) float32 in Square meter
-func AcretoSquaremeter(acre float32) float32 {
+// AcretoSquaremeter (acre float64) float64 in Square meter
+func AcretoSquaremeter(acre float64) float64 {
 	return acre / 4046.856
 }
 
-// AcretoSquarekilometer (acre float32) float32 in Square mile
-func AcretoSquarekilometer(acre float32) float32 {
+// AcretoSquarekilometer (acre float64) float64 in Square mile
+func AcretoSquarekilometer(acre float64) float64 {
 	return acre / 247.105
 }
 
-// HectaretoAcre (hectare float32) float32 in Acre value
-func HectaretoAcre(hectare float32) float32 {
+// HectaretoAcre (hectare float64) float64 in Acre value
+func HectaretoAcre(hectare float64) float64 {
 	return hectare * 2.471
 }
 
-// HectaretoSquareinch (hectare float32) float32 in Square Inch value
-func HectaretoSquareinch(hectare float32) float32 {
+// HectaretoSquareinch (hectare float64) float64 in Square Inch value
+func HectaretoSquareinch(hectare float64) float64 {
 	return hectare * 1.55e+7
 }
 
-// HectaretoSquarefoot (hectare float32) float32 in Square Foot value
-func HectaretoSquarefoot(hectare float32) float32 {
+// HectaretoSquarefoot (hectare float64) float64 in Square Foot value
+func HectaretoSquarefoot(hectare float64) float64 {
 	return hectare * 107639.104
 }
 
-// HectaretoSquareyard (hectare float32) float32 in Square yard value
-func HectaretoSquareyard(hectare float32) float32 {
+// HectaretoSquareyard (hectare float64) float64 in Square yard value
+func HectaretoSquareyard(hectare float64) float64 {
 	return hectare * 107639.104
 }
 
-// HectaretoSquaremile (hectare float32) float32 in Square mile value
-func HectaretoSquaremile(hectare float32) float32 {
+// HectaretoSquaremile (hectare float64) float64 in Square mile value
+func HectaretoSquaremile(hectare float64) float64 {
 	return hectare / 258.999
 }
 
-// HectaretoSquaremeter (hectare float32) float32 in Square meter value
-func HectaretoSquaremeter(hectare float32) float32 {
+// HectaretoSquaremeter (hectare float64) float64 in Square meter value
+func HectaretoSquaremeter(hectare float64) float64 {
 	return hectare * 10000
 }
 
-// HectaretoSquarekilometer (hectare float32) float32 in Square kilometer value
-func HectaretoSquarekilometer(hectare float32) float32 {
+// HectaretoSquarekilometer (hectare float64) float64 in Square kilometer value
+func HectaretoSquarekilometer(hectare float64) float64 {
 	return hectare / 100
 }
 
-// SquareinchtoAcre (squareinch float32) float32 in Acre value
-func SquareinchtoAcre(squareinch float32) float32 {
+// SquareinchtoAcre (squareinch float64) float64 in Acre value
+func SquareinchtoAcre(squareinch float64) float64 {
 	return squareinch / 6.273e+6
 }
 
-// SquareinchtoHectare (squareinch float32) float32 in hectare value
-func SquareinchtoHectare(squareinch float32) float32 {
+// SquareinchtoHectare (squareinch float64) float64 in hectare value
+func SquareinchtoHectare(squareinch float64) float64 {
 	return squareinch / 1.55e+7
 }
 
-// SquareinchtoSquarefoot (squarefoot float32) float32 in Square Foot value
-func SquareinchtoSquarefoot(squareinch float32) float32 {
+// SquareinchtoSquarefoot (squarefoot float64) float64 in Square Foot value
+func SquareinchtoSquarefoot(squareinch float64) float64 {
 	return squareinch / 144
 }
 
-// SquareinchtoSquareyard (squareinch float32) float32 in square yard value
-func SquareinchtoSquareyard(squareinch float32) float32 {
+// SquareinchtoSquareyard (squareinch float64) float64 in square yard value
+func SquareinchtoSquareyard(squareinch float64) float64 {
 	return squareinch / 1296
 }
 
-// SquareinchtoSquaremile (squareinch float32) float32 Square mile
-func SquareinchtoSquaremile(squareinch float32) float32 {
+// SquareinchtoSquaremile (squareinch float64) float64 Square mile
+func SquareinchtoSquaremile(squareinch float64) float64 {
 	return squareinch / 4.014e+9
 }
 
-// SquareinchtoSquaremeter (squareinch float32) float32 Square meter value
-func SquareinchtoSquaremeter(squareinch float32) float32 {
+// SquareinchtoSquaremeter (squareinch float64) float64 Square meter value
+func SquareinchtoSquaremeter(squareinch float64) float64 {
 	return squareinch / 1550.003
 }
 
-// SquareinchtoSquarekilometer (squareinch float32) float32 Square kilometer value
-func SquareinchtoSquarekilometer(squareinch float32) float32 {
+// SquareinchtoSquarekilometer (squareinch float64) float64 Square kilometer value
+func SquareinchtoSquarekilometer(squareinch float64) float64 {
 	return squareinch / 1.55e+9
 }
 
-// SquarefoottoAcre (squareinch float32) float32 Square Acre value
-func SquarefoottoAcre(squarefoot float32) float32 {
+// SquarefoottoAcre (squareinch float64) float64 Square Acre value
+func SquarefoottoAcre(squarefoot float64) float64 {
 	return squarefoot / 43560
 }
 
-// SquarefoottoHectare (squareinch float32) float32 Square hectare value
-func SquarefoottoHectare(squarefoot float32) float32 {
+// SquarefoottoHectare (squareinch float64) float64 Square hectare value
+func SquarefoottoHectare(squarefoot float64) float64 {
 	return squarefoot / 107639.104
 }
 
-// SquarefoottoSquareinch (squareinch float32) float32 Square Square inch value
-func SquarefoottoSquareinch(squarefoot float32) float32 {
+// SquarefoottoSquareinch (squareinch float64) float64 Square Square inch value
+func SquarefoottoSquareinch(squarefoot float64) float64 {
 	return squarefoot * 144
 }
 
-// SquarefoottoSquareyard (squareinch float32) float32 Square Square yard value
-func SquarefoottoSquareyard(squarefoot float32) float32 {
+// SquarefoottoSquareyard (squareinch float64) float64 Square Square yard value
+func SquarefoottoSquareyard(squarefoot float64) float64 {
 	return squarefoot / 1296
 }
 
-// SquarefoottoSquaremile (squareinch float32) float32 Square Square mile value
-func SquarefoottoSquaremile(squarefoot float32) float32 {
+// SquarefoottoSquaremile (squareinch float64) float64 Square Square mile value
+func SquarefoottoSquaremile(squarefoot float64) float64 {
 	return squarefoot / 4.014e+9
 }
 
-// SquarefoottoSquaremeter (squareinch float32) float32 Square Square meter value
-func SquarefoottoSquaremeter(squarefoot float32) float32 {
+// SquarefoottoSquaremeter (squareinch float64) float64 Square Square meter value
+func SquarefoottoSquaremeter(squarefoot float64) float64 {
 	return squarefoot / 1550.003
 }
 
-// SquarefoottoSquarekilometer (squareinch float32) float32 Square Square kilometer value
-func SquarefoottoSquarekilometer(squarefoot float32) float32 {
+// SquarefoottoSquarekilometer (squareinch float64) float64 Square Square kilometer value
+func SquarefoottoSquarekilometer(squarefoot float64) float64 {
 	return squarefoot / 1.55e+9
 }
 
-// SquareyardtoAcre (squareyard float32) float32 acre in value
-func SquareyardtoAcre(squareyard float32) float32 {
+// SquareyardtoAcre (squareyard float64) float64 acre in value
+func SquareyardtoAcre(squareyard float64) float64 {
 	return squareyard / 4840
 }
 
-// SquareyardtoHectare (squareyard float32) float32 acre in Hectare value
-func SquareyardtoHectare(squareyard float32) float32 {
+// SquareyardtoHectare (squareyard float64) float64 acre in Hectare value
+func SquareyardtoHectare(squareyard float64) float64 {
 	return squareyard / 11959.9
 }
 
-// SquareyardtoSquareinch (squareyard float32) float32 acre in Square inch value
-func SquareyardtoSquareinch(squareyard float32) float32 {
+// SquareyardtoSquareinch (squareyard float64) float64 acre in Square inch value
+func SquareyardtoSquareinch(squareyard float64) float64 {
 	return squareyard * 1296
 }
 
-// SquareyardtoSquarefoot (squareyard float32) float32 acre in Square foot value
-func SquareyardtoSquarefoot(squareyard float32) float32 {
+// SquareyardtoSquarefoot (squareyard float64) float64 acre in Square foot value
+func SquareyardtoSquarefoot(squareyard float64) float64 {
 	return squareyard * 9
 }
 
-// SquareyardtoSquaremile (squareyard float32) float32 acre in Square mile value
-func SquareyardtoSquaremile(squareyard float32) float32 {
+// SquareyardtoSquaremile (squareyard float64) float64 acre in Square mile value
+func SquareyardtoSquaremile(squareyard float64) float64 {
 	return squareyard / 3.098e+6
 }
 
-// SquareyardtoSquaremeter (squareyard float32) float32 acre in Square meter value
-func SquareyardtoSquaremeter(squareyard float32) float32 {
+// SquareyardtoSquaremeter (squareyard float64) float64 acre in Square meter value
+func SquareyardtoSquaremeter(squareyard float64) float64 {
 	return squareyard / 1.196
 }
 
-// SquareyardtoSquarekilometer (squareyard float32) float32 acre in Square kilometer value
-func SquareyardtoSquarekilometer(squareyard float32) float32 {
+// SquareyardtoSquarekilometer (squareyard float64) float64 acre in Square kilometer value
+func SquareyardtoSquarekilometer(squareyard float64) float64 {
 	return squareyard / 1.196e+6
 }
 
-//SquaremiletoAcre (squaremile float32) float32
-func SquaremiletoAcre(squaremile float32) float32 {
+//SquaremiletoAcre (squaremile float64) float64
+func SquaremiletoAcre(squaremile float64) float64 {
 	return squaremile * 640
 }
 
-//SquaremiletoHectare (squaremile float32) float32
-func SquaremiletoHectare(squaremile float32) float32 {
+//SquaremiletoHectare (squaremile float64) float64
+func SquaremiletoHectare(squaremile float64) float64 {
 	return squaremile * 258.999
 }
 
-//SquaremiletoSquareinch (squaremile float32) float32
-func SquaremiletoSquareinch(squaremile float32) float32 {
+//SquaremiletoSquareinch (squaremile float64) float64
+func SquaremiletoSquareinch(squaremile float64) float64 {
 	return squaremile * 4.014e+9
 }
 
-//SquaremiletoSquarefoot (squaremile float32) float32
-func SquaremiletoSquarefoot(squaremile float32) float32 {
+//SquaremiletoSquarefoot (squaremile float64) float64
+func SquaremiletoSquarefoot(squaremile float64) float64 {
 	return squaremile * 2.788e+7
 }
 
-// SquaremiletoSquareyard (squaremile float32) float32
-func SquaremiletoSquareyard(squaremile float32) float32 {
+// SquaremiletoSquareyard (squaremile float64) float64
+func SquaremiletoSquareyard(squaremile float64) float64 {
 	return squaremile * 3.098e+6
 }
 
-//Squaremiletometer (squaremile float32) float32
-func Squaremiletometer(squaremile float32) float32 {
+//Squaremiletometer (squaremile float64) float64
+func Squaremiletometer(squaremile float64) float64 {
 	return squaremile * 2.59e+6
 }
 
-//Squaremiletokilometer (squaremile float32) float32
-func Squaremiletokilometer(squaremile float32) float32 {
+//Squaremiletokilometer (squaremile float64) float64
+func Squaremiletokilometer(squaremile float64) float64 {
 	return squaremile * 2.59
 }
 
-// SquarekilometertoAcre (squarekilometer float32) float32
-func SquarekilometertoAcre(squarekilometer float32) float32 {
+// SquarekilometertoAcre (squarekilometer float64) float64
+func SquarekilometertoAcre(squarekilometer float64) float64 {
 	return squarekilometer * 247.105
 }
 
-// SquarekilometertoHectare (squarekilometer float32) float32
-func SquarekilometertoHectare(squarekilometer float32) float32 {
+// SquarekilometertoHectare (squarekilometer float64) float64
+func SquarekilometertoHectare(squarekilometer float64) float64 {
 	return squarekilometer * 100
 }
 
-// SquarekilometertoSquareinch (squarekilometer float32) float32
-func SquarekilometertoSquareinch(squarekilometer float32) float32 {
+// SquarekilometertoSquareinch (squarekilometer float64) float64
+func SquarekilometertoSquareinch(squarekilometer float64) float64 {
 	return squarekilometer * 1.55e+9
 }
 
-// SquarekilometertoSquarefoot (squarekilometer float32) float32
-func SquarekilometertoSquarefoot(squarekilometer float32) float32 {
+// SquarekilometertoSquarefoot (squarekilometer float64) float64
+func SquarekilometertoSquarefoot(squarekilometer float64) float64 {
 	return squarekilometer * 1.076e+7
 }
 
-// SquarekilometertoSquareyard (squarekilometer float32) float32
-func SquarekilometertoSquareyard(squarekilometer float32) float32 {
+// SquarekilometertoSquareyard (squarekilometer float64) float64
+func SquarekilometertoSquareyard(squarekilometer float64) float64 {
 	return squarekilometer * 1.196e+6
 }
 
-// SquarekilometertoSquaremile (squarekilometer float32) float32
-func SquarekilometertoSquaremile(squarekilometer float32) float32 {
+// SquarekilometertoSquaremile (squarekilometer float64) float64
+func SquarekilometertoSquaremile(squarekilometer float64) float64 {
 	return squarekilometer / 2.59
 }
 
-// SquarekilometertoSquaremeter (squarekilometer float32) float32
-func SquarekilometertoSquaremeter(squarekilometer float32) float32 {
+// SquarekilometertoSquaremeter (squarekilometer float64) float64
+func SquarekilometertoSquaremeter(squarekilometer float64) float64 {
 	return squarekilometer * 1e+6
 }
